Add concepts section (#293)
In an attempt to make it easier for people to understand what data
sources OpenTelemetry supports as well as the various components, I
added a concepts section. In addition, created a glossary within this
new section (addresses #193). The idea was to provide basic information
and link to authoritative resources as much as possible.

CREDIT: The information included pulls from the prior work of other
including:
- OpenTelemetry READMEs
- OpenCensus documentation
- Kubernetes documentation structure