Otel Icons and zipfile (#333)
* Otel Icons and zipfile

This is a PR to submit the icons and zipfile. I tried to clean it up. We only have .png for  512x512, the 32x32 are svg but I can convert to png if we want. The zipfile contains the same content as the directory.

* New components.md clarifying collector builds

I have added a paragraph to clarify the two builds for the collector and why they both exist.

* Revert "New components.md clarifying collector builds"

This reverts commit 58b55a7df456d2d75fab9cbb7778bd3de1e1183a.