Update new-end-user-resources.md
Adding a quick section about the talks to this post so that everything is a complete circle and we can easily link here to represent the WG's full activities. The section is copied directly from the other blog post. I'm experimenting with doing this directly in github rather than gitpod so let's see how it goes!