Make span kind description closer the the spec
Spec language is soft in regards of clients/server span relationships:

https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/api.md#spankind

Server span does not need to have a parent, it could also be that multiple server implementations that are not aware of each other create server->server spans. While not ideal, this is a reality.

Client spans can be nested under other client spans. 
E.g.  DB ORM creates a client span and the underlying DB driver creates a client span. They are unaware of each other and have no knowledge if they are both enabled in the same application.