Update vendors.yaml to add ClickHouse
Hi! Adding ClickHouse to OpenTelemetry vendor list, per the vendor docs instructions.

Link to the documentation that details how your offering consumes OpenTelemetry natively via OTLP.
https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/exporter/clickhouseexporter
https://clickhouse.com/blog/storing-log-data-in-clickhouse-fluent-bit-vector-open-telemetry
https://clickhouse.com/blog/storing-traces-and-spans-open-telemetry-in-clickhouse
https://clickhouse.com/use-cases/logging-and-metrics

Link to your distribution, if applicable
N/A

Link that proves that your offering is open source, if applicable. An open source distribution does not qualify your offering to be marked “open source”.
https://github.com/ClickHouse/ClickHouse/blob/master/LICENSE

GitHub handle or email address as a point of contact so that we can reach out in case we have questions
@tbragin