Update content/en/blog/2024/otel-cicd-sig/index.md
per @tiffany76 review:

Although trace and event are defined terms in OpenTelemetry, they are common nouns and should not be capitalized. To illustrate: the OpenTelemetry Collector is capitalized because it is a proper noun referring to a specific collector. If you refer to other collectors, the term is not capitalized.

Co-authored-by: Tiffany Hrabusa <30397949+tiffany76@users.noreply.github.com>