Provide full example collector configuration of using the load-balancing exporter, tail-sampling processor, and span metrics connector together when scaled to multiple collector instances.
Also removes language and configuration suggesting that load-balancing should be a different deployment of collectors than the collectors doing the tail sampling and span metric generation. It's easier to maintain a single deployment responsible for both load balancing and processing of the load balanced data but the pattern for doing this may not be obvious at first.