JS instrumentation/propagation docs: switch to tsx for on-the-fly TypeScript compilation
Similar to https://github.com/open-telemetry/opentelemetry.io/pull/7606
this updates the remaining parts of the JavaScript tracing docs to use
'tsx' rather than the no-longer-maintained 'ts-node' for TypeScript
examples.

As well (as in #7606) this updates to '--import' usage over '--require'
for preloading OTel bootstrap code. This is in prep for future ESM
support, and helps with some confusion around automatic type-stripping
in recent versions of Node.js.

Refs: https://github.com/open-telemetry/opentelemetry.io/issues/7623